+++ packages/common/src/types/dbt.ts
@@ -23,6 +23,7 @@ export enum SupportedDbtAdapter {
     SNOWFLAKE = 'snowflake',
     REDSHIFT = 'redshift',
     POSTGRES = 'postgres',
+    TRINO = 'trino',
 }
 
 export type DbtNodeConfig = {
@@ -104,6 +105,7 @@ export const normaliseModelDatabase = (
         case SupportedDbtAdapter.POSTGRES:
         case SupportedDbtAdapter.BIGQUERY:
         case SupportedDbtAdapter.SNOWFLAKE:
+        case SupportedDbtAdapter.TRINO:
         case SupportedDbtAdapter.REDSHIFT:
             if (model.database === null) {
                 throw new ParseError(