+++ packages/frontend/src/components/DashboardTiles/TileBase/index.tsx
@@ -98,7 +98,7 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
                                         <>
                                             <MenuItem2
                                                 icon="edit"
-                                                text="Edit tile"
+                                                text="Replace chart"
                                                 onClick={() =>
                                                     setIsEditing(true)
                                                 }