+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## [0.373.1](https://github.com/lightdash/lightdash/compare/0.373.0...0.373.1) (2023-01-17)
+
+
+### Bug Fixes
+
+* override default props for blueprint dialog - do not close on escape or outside click ([#4175](https://github.com/lightdash/lightdash/issues/4175)) ([1682b71](https://github.com/lightdash/lightdash/commit/1682b7108f3b4fade015303aa0456ba5f98c2d63))
+
+# [0.373.0](https://github.com/lightdash/lightdash/compare/0.372.0...0.373.0) (2023-01-17)
+
+
+### Features
+
+* conditional formatting multiple conditions ([#4164](https://github.com/lightdash/lightdash/issues/4164)) ([b030323](https://github.com/lightdash/lightdash/commit/b03032330e66a71d6430090ec22a8ee4eadf1821))
+
+# [0.372.0](https://github.com/lightdash/lightdash/compare/0.371.1...0.372.0) (2023-01-17)
+
+
+### Bug Fixes
+
+* add filter input width ([#4151](https://github.com/lightdash/lightdash/issues/4151)) ([eea61b3](https://github.com/lightdash/lightdash/commit/eea61b33dcb6ca4d2e8c4514af697024e33a27c4))
+* rename 'edit tile' menu option ([#4188](https://github.com/lightdash/lightdash/issues/4188)) ([c588d65](https://github.com/lightdash/lightdash/commit/c588d65af6771fd7416a3506ea52526e19f0f89a))
+
+
+### Features
+
+* support for Trino as datasource ([#4101](https://github.com/lightdash/lightdash/issues/4101)) ([a453a5a](https://github.com/lightdash/lightdash/commit/a453a5a5f988bc7bfcd67313507f8fb5b49e9b0d))
+
+## [0.371.1](https://github.com/lightdash/lightdash/compare/0.371.0...0.371.1) (2023-01-17)
+
+
+### Bug Fixes
+
+* numeric input should allow negative values ([#4176](https://github.com/lightdash/lightdash/issues/4176)) ([dca110f](https://github.com/lightdash/lightdash/commit/dca110fcf299c06bbce6266e8089bb0e13820e2f))
+
 # [0.371.0](https://github.com/lightdash/lightdash/compare/0.370.4...0.371.0) (2023-01-16)